@@ -92,20 +92,20 @@ public class SchoolGalleryFragment extends Fragment {
     }
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         View view =  inflater.inflate(R.layout.fragment_school_gallery, container, false);
-        ConnectionUtil.checkConnection(getActivity().findViewById(R.id.placeSnackBar));
+        if(getActivity() !=null) ConnectionUtil.checkConnection(getActivity().findViewById(R.id.placeSnackBar));
 
-        final LinearLayout not_found = (LinearLayout) view.findViewById(R.id.not_found);
-        final TextView not_found_text1 = (TextView) view.findViewById(R.id.not_found_text1);
-        final TextView not_found_text2 = (TextView) view.findViewById(R.id.not_found_text2);
+        final LinearLayout not_found = view.findViewById(R.id.not_found);
+        final TextView not_found_text1 = view.findViewById(R.id.not_found_text1);
+        final TextView not_found_text2 = view.findViewById(R.id.not_found_text2);
 
         not_found_text1.setText("No Photos For this place!");
         not_found_text2.setVisibility(View.GONE);
 
-        recyclerView = (RecyclerView) view.findViewById(R.id.recycler_view);
-        recycler_view_new = (RecyclerView) view.findViewById(R.id.recycler_view_new);
+        recyclerView =  view.findViewById(R.id.recycler_view);
+        recycler_view_new = view.findViewById(R.id.recycler_view_new);
         if (images.size()==0){
             not_found.setVisibility(View.VISIBLE);
             recyclerView.setVisibility(View.GONE);
@@ -135,78 +135,62 @@ public class SchoolGalleryFragment extends Fragment {
 
         hmap = new TreeMap<>();
         mProgressDialog = new ProgressDialog(getActivity());
-        linearLayoutEdit = (LinearLayout) view.findViewById(R.id.linearLayoutEdit);
+        linearLayoutEdit = view.findViewById(R.id.linearLayoutEdit);
         FirebaseAuth auth = FirebaseAuth.getInstance();
         FirebaseUser user = auth.getCurrentUser();
         if (user != null && user.getUid().equals(uid)){
             linearLayoutEdit.setVisibility(View.VISIBLE);
-            Button add_photos_school = (Button) view.findViewById(R.id.add_photos_school);
-            add_photos_school.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    Intent intent = new Intent(getActivity(), AlbumSelectActivity.class);
-                    intent.putExtra(ConstantsCustomGallery.INTENT_EXTRA_LIMIT, 10); // set limit for image selection
-                    startActivityForResult(intent, ConstantsCustomGallery.REQUEST_CODE);
-                }
+            Button add_photos_school = view.findViewById(R.id.add_photos_school);
+            add_photos_school.setOnClickListener(v -> {
+                Intent intent = new Intent(getActivity(), AlbumSelectActivity.class);
+                intent.putExtra(ConstantsCustomGallery.INTENT_EXTRA_LIMIT, 10); // set limit for image selection
+                startActivityForResult(intent, ConstantsCustomGallery.REQUEST_CODE);
             });
 
-            Button btnClearImages = (Button) view.findViewById(R.id.btnClearImages);
-            btnClearImages.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    imagesNew.clear();
-                    mAdapterNew.notifyDataSetChanged();
-                }
+            Button btnClearImages = view.findViewById(R.id.btnClearImages);
+            btnClearImages.setOnClickListener(v -> {
+                imagesNew.clear();
+                mAdapterNew.notifyDataSetChanged();
             });
 
-            Button btnSaveImages = (Button) view.findViewById(R.id.btnSaveImages);
-            btnSaveImages.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    if (hmap.size()>0) {
-                        mProgressDialog.setMessage("Uploading Photos");
-                        mProgressDialog.setProgressStyle(ProgressDialog.STYLE_SPINNER);
-                        mProgressDialog.setIndeterminate(true);
-                        mProgressDialog.setCancelable(false);
-                        mProgressDialog.show();
-                        final DatabaseReference databaseReference = FirebaseDatabase.getInstance().getReference().child("schools").child(schoolId);
-                        StorageReference mStorageReference = FirebaseStorage.getInstance().getReference().child("school_photos");
-                        StorageReference filepath;
-                        for (final Map.Entry entry : hmap.entrySet()) {
-                            Uri uri = (Uri) entry.getValue();
-                            filepath = mStorageReference.child(schoolId).child(uri.getLastPathSegment() + "-" + new Timestamp(System.currentTimeMillis()));
-                            filepath.putFile(uri).addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
-                                @Override
-                                public void onSuccess(UploadTask.TaskSnapshot taskSnapshot) {
-                                    @SuppressWarnings("VisibleForTests") Uri downloadURI = taskSnapshot.getDownloadUrl();
-                                    databaseReference.child("images").push().setValue(downloadURI.toString());
+            Button btnSaveImages = view.findViewById(R.id.btnSaveImages);
+            btnSaveImages.setOnClickListener(v -> {
+                if (hmap.size()>0) {
+                    mProgressDialog.setMessage("Uploading Photos");
+                    mProgressDialog.setProgressStyle(ProgressDialog.STYLE_SPINNER);
+                    mProgressDialog.setIndeterminate(true);
+                    mProgressDialog.setCancelable(false);
+                    mProgressDialog.show();
+                    final DatabaseReference databaseReference = FirebaseDatabase.getInstance().getReference().child("schools").child(schoolId);
+                    StorageReference mStorageReference = FirebaseStorage.getInstance().getReference().child("school_photos");
+                    StorageReference filepath;
+                    for (final Map.Entry entry : hmap.entrySet()) {
+                        Uri uri = (Uri) entry.getValue();
+                        filepath = mStorageReference.child(schoolId).child(uri.getLastPathSegment() + "-" + new Timestamp(System.currentTimeMillis()));
+                        filepath.putFile(uri).addOnSuccessListener(taskSnapshot -> {
+                            @SuppressWarnings("VisibleForTests") Uri downloadURI = taskSnapshot.getDownloadUrl();
+                            databaseReference.child("images").push().setValue(downloadURI.toString());
 
-                                    if (entry.equals(hmap.lastEntry())) {
-                                        mProgressDialog.dismiss();
-                                        AlertDialog.Builder builder = new AlertDialog.Builder(
-                                                getActivity());
-                                        builder.setCancelable(true);
-                                        builder.setMessage("Congrats, your photos posted successfully ! Once we review it we will post it live.");
-                                        builder.setPositiveButton("Ok",
-                                                new DialogInterface.OnClickListener() {
-                                                    @Override
-                                                    public void onClick(DialogInterface dialog,
-                                                                        int which) {
-                                                        imagesNew.clear();
-                                                        mAdapterNew.notifyDataSetChanged();
-                                                        dialog.dismiss();
-                                                    }
-                                                });
-                                        AlertDialog alert = builder.create();
-                                        alert.show();
-                                    }
-                                }
-                            });
-                        }
-                    }else {
-                        new CustomToast().Show_Toast(getActivity(),
-                                "Please add atleast one photo to save.");
+                            if (entry.equals(hmap.lastEntry())) {
+                                mProgressDialog.dismiss();
+                                AlertDialog.Builder builder = new AlertDialog.Builder(
+                                        getActivity());
+                                builder.setCancelable(true);
+                                builder.setMessage("Congrats, your photos posted successfully ! Once we review it we will post it live.");
+                                builder.setPositiveButton("Ok",
+                                        (dialog, which) -> {
+                                            imagesNew.clear();
+                                            mAdapterNew.notifyDataSetChanged();
+                                            dialog.dismiss();
+                                        });
+                                AlertDialog alert = builder.create();
+                                alert.show();
+                            }
+                        });
                     }
+                }else {
+                    new CustomToast().Show_Toast(getActivity(),
+                            "Please add atleast one photo to save.");
                 }
             });
         }
@@ -264,6 +248,6 @@ public class SchoolGalleryFragment extends Fragment {
     @Override
     public void onResume() {
         super.onResume();
-        ConnectionUtil.checkConnection(getActivity().findViewById(R.id.placeSnackBar));
+        if(getActivity() !=null) ConnectionUtil.checkConnection(getActivity().findViewById(R.id.placeSnackBar));
     }
 }
