@@ -3,6 +3,7 @@ package com.credolabs.justcredo.school;
 
 import android.content.DialogInterface;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AlertDialog;
@@ -67,18 +68,18 @@ public class SchoolGenreFragment extends Fragment {
     }
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_school_genre, container, false);
-        progress = (ProgressBar) view.findViewById(R.id.progress);
-        final LinearLayout not_found = (LinearLayout) view.findViewById(R.id.not_found);
-        final TextView not_found_text1 = (TextView) view.findViewById(R.id.not_found_text1);
-        final TextView not_found_text2 = (TextView) view.findViewById(R.id.not_found_text2);
+        progress = view.findViewById(R.id.progress);
+        final LinearLayout not_found = view.findViewById(R.id.not_found);
+        final TextView not_found_text1 = view.findViewById(R.id.not_found_text1);
+        final TextView not_found_text2 = view.findViewById(R.id.not_found_text2);
 
         not_found_text1.setText("No genres added yet !");
         not_found_text2.setVisibility(View.GONE);
         dance_section = (LinearLayout) view.findViewById(R.id.dance_section);
-        ExpandableHeightGridView dance_gridview = (ExpandableHeightGridView) view.findViewById(R.id.dance_gridview);
+        ExpandableHeightGridView dance_gridview = view.findViewById(R.id.dance_gridview);
         if (model.getDancing() != null) {
             not_found.setVisibility(View.GONE);
             ArrayList<String> list = new ArrayList<>(model.getDancing().values());
@@ -88,8 +89,8 @@ public class SchoolGenreFragment extends Fragment {
             dance_section.setVisibility(View.GONE);
         }
 
-        singing_section = (LinearLayout) view.findViewById(R.id.singing_section);
-        ExpandableHeightGridView expandableHeightGridView = (ExpandableHeightGridView) view.findViewById(R.id.singing_gridview);
+        singing_section = view.findViewById(R.id.singing_section);
+        ExpandableHeightGridView expandableHeightGridView = view.findViewById(R.id.singing_gridview);
         if (model.getSinging() != null) {
             not_found.setVisibility(View.GONE);
             ArrayList<String> classList = new ArrayList<>(model.getSinging().values());
@@ -99,8 +100,8 @@ public class SchoolGenreFragment extends Fragment {
             singing_section.setVisibility(View.GONE);
         }
 
-        instruments_section = (LinearLayout) view.findViewById(R.id.instruments_section);
-        ExpandableHeightGridView linear_layout_curriculumn = (ExpandableHeightGridView) view.findViewById(R.id.instruments_gridview);
+        instruments_section = view.findViewById(R.id.instruments_section);
+        ExpandableHeightGridView linear_layout_curriculumn = view.findViewById(R.id.instruments_gridview);
         if (model.getInstruments() != null) {
             not_found.setVisibility(View.GONE);
             ArrayList<String> classList = new ArrayList<>(model.getInstruments().values());
@@ -110,8 +111,8 @@ public class SchoolGenreFragment extends Fragment {
             instruments_section.setVisibility(View.GONE);
         }
 
-        other_genres_section = (LinearLayout) view.findViewById(R.id.other_genres_section);
-        ExpandableHeightGridView other_genres_gridview = (ExpandableHeightGridView) view.findViewById(R.id.other_genres_gridview);
+        other_genres_section = view.findViewById(R.id.other_genres_section);
+        ExpandableHeightGridView other_genres_gridview = view.findViewById(R.id.other_genres_gridview);
         if (model.getOther_genres() != null) {
             not_found.setVisibility(View.GONE);
             ArrayList<String> classList = new ArrayList<>(model.getOther_genres().values());
@@ -122,95 +123,75 @@ public class SchoolGenreFragment extends Fragment {
         }
 
         if (Util.checkSchoolAdmin(model)) { // edit section only for admin
-            fragContainer = (LinearLayout) view.findViewById(R.id.container);
+            fragContainer = view.findViewById(R.id.container);
 
-            save_cancel = (LinearLayout) view.findViewById(R.id.save_cancel);
-            LinearLayout edit_music_section = (LinearLayout) view.findViewById(R.id.edit_music_section);
-            Button cancel_music = (Button) view.findViewById(R.id.cancel_music);
-            edit_music = (Button) view.findViewById(R.id.edit_music);
-            Button save_music = (Button) view.findViewById(R.id.save_music);
+            save_cancel = view.findViewById(R.id.save_cancel);
+            LinearLayout edit_music_section = view.findViewById(R.id.edit_music_section);
+            Button cancel_music = view.findViewById(R.id.cancel_music);
+            edit_music = view.findViewById(R.id.edit_music);
+            Button save_music = view.findViewById(R.id.save_music);
 
             edit_music_section.setVisibility(View.VISIBLE);
             final GenresFillFragment genresFillFragment = GenresFillFragment.newInstance(
                     PlaceTypes.Action.EDIT_BACKUP.getValue(), model);
-            edit_music.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    progress.setVisibility(View.VISIBLE);
-                    not_found.setVisibility(View.GONE);
-                    FragmentTransaction transaction = getChildFragmentManager().beginTransaction();
-                    transaction.replace(R.id.container, genresFillFragment);
-                    transaction.addToBackStack(null);
-                    transaction.commit();
-                    hideContent();
-                    fragContainer.setVisibility(View.VISIBLE);
-                    edit_music.setVisibility(View.GONE);
-                    save_cancel.setVisibility(View.VISIBLE);
-                    progress.setVisibility(View.GONE);
-                }
+            edit_music.setOnClickListener(v -> {
+                progress.setVisibility(View.VISIBLE);
+                not_found.setVisibility(View.GONE);
+                FragmentTransaction transaction = getChildFragmentManager().beginTransaction();
+                transaction.replace(R.id.container, genresFillFragment);
+                transaction.addToBackStack(null);
+                transaction.commit();
+                hideContent();
+                fragContainer.setVisibility(View.VISIBLE);
+                edit_music.setVisibility(View.GONE);
+                save_cancel.setVisibility(View.VISIBLE);
+                progress.setVisibility(View.GONE);
             });
 
-            cancel_music.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    fragContainer.setVisibility(View.GONE);
-                    edit_music.setVisibility(View.VISIBLE);
-                    save_cancel.setVisibility(View.GONE);
-                    visibleContent();
-                }
+            cancel_music.setOnClickListener(v -> {
+                fragContainer.setVisibility(View.GONE);
+                edit_music.setVisibility(View.VISIBLE);
+                save_cancel.setVisibility(View.GONE);
+                visibleContent();
             });
 
-            save_music.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(getActivity());
-                    alertDialogBuilder.setMessage("Do you want to update genres ?");
-                    alertDialogBuilder.setPositiveButton("Yes",
-                            new DialogInterface.OnClickListener() {
-                                @Override
-                                public void onClick(DialogInterface arg0, int arg1) {
-                                    progress.setVisibility(View.VISIBLE);
-                                    HashMap<String, HashMap<String, String>> map = genresFillFragment.getFragmentState();
-                                    DatabaseReference mDatabaseReference = FirebaseDatabase.getInstance().getReference().child(School.SCHOOL_DATABASE);
+            save_music.setOnClickListener(v -> {
+                AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(getActivity());
+                alertDialogBuilder.setMessage("Do you want to update genres ?");
+                alertDialogBuilder.setPositiveButton("Yes",
+                        new DialogInterface.OnClickListener() {
+                            @Override
+                            public void onClick(DialogInterface arg0, int arg1) {
+                                progress.setVisibility(View.VISIBLE);
+                                HashMap<String, HashMap<String, String>> map = genresFillFragment.getFragmentState();
+                                DatabaseReference mDatabaseReference = FirebaseDatabase.getInstance().getReference().child(School.SCHOOL_DATABASE);
 
-                                    mDatabaseReference.child(model.getId()).child(School.DANCING).setValue(map.get(School.DANCING));
-                                    mDatabaseReference.child(model.getId()).child(School.SINGING).setValue(map.get(School.SINGING));
-                                    mDatabaseReference.child(model.getId()).child(School.INSTRUMENTS).setValue(map.get(School.INSTRUMENTS));
-                                    mDatabaseReference.child(model.getId()).child(School.OTHER_GENRES).setValue(map.get(School.OTHER_GENRES));
-                                    fragContainer.setVisibility(View.GONE);
-                                    edit_music.setVisibility(View.VISIBLE);
-                                    save_cancel.setVisibility(View.GONE);
-                                    visibleContent();
-                                    progress.setVisibility(View.GONE);
-                                    AlertDialog.Builder builder = new AlertDialog.Builder(
-                                            getActivity());
-                                    builder.setCancelable(true);
-                                    builder.setMessage("Congrats, genres updated successfully ! You can write blog and upload photos so that users" +
-                                            " can know about this.");
-                                    builder.setPositiveButton("Ok",
-                                            new DialogInterface.OnClickListener() {
-                                                @Override
-                                                public void onClick(DialogInterface dialog,
-                                                                    int which) {
-                                                    dialog.dismiss();
-                                                }
-                                            });
-                                    AlertDialog alert = builder.create();
-                                    alert.show();
-                                }
-                            });
+                                mDatabaseReference.child(model.getId()).child(School.DANCING).setValue(map.get(School.DANCING));
+                                mDatabaseReference.child(model.getId()).child(School.SINGING).setValue(map.get(School.SINGING));
+                                mDatabaseReference.child(model.getId()).child(School.INSTRUMENTS).setValue(map.get(School.INSTRUMENTS));
+                                mDatabaseReference.child(model.getId()).child(School.OTHER_GENRES).setValue(map.get(School.OTHER_GENRES));
+                                fragContainer.setVisibility(View.GONE);
+                                edit_music.setVisibility(View.VISIBLE);
+                                save_cancel.setVisibility(View.GONE);
+                                visibleContent();
+                                progress.setVisibility(View.GONE);
+                                AlertDialog.Builder builder = new AlertDialog.Builder(
+                                        getActivity());
+                                builder.setCancelable(true);
+                                builder.setMessage("Congrats, genres updated successfully ! You can write blog and upload photos so that users" +
+                                        " can know about this.");
+                                builder.setPositiveButton("Ok",
+                                        (dialog, which) -> dialog.dismiss());
+                                AlertDialog alert = builder.create();
+                                alert.show();
+                            }
+                        });
 
-                    alertDialogBuilder.setNegativeButton("No", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            dialog.dismiss();
-                        }
-                    });
+                alertDialogBuilder.setNegativeButton("No", (dialog, which) -> dialog.dismiss());
 
-                    alertDialogBuilder.create();
-                    alertDialogBuilder.show();
+                alertDialogBuilder.create();
+                alertDialogBuilder.show();
 
-                }
             });
         }
         return view;
