@@ -3,15 +3,13 @@ package com.credolabs.justcredo.school;
 import android.app.Activity;
 import android.app.ProgressDialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
-import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Handler;
+import android.support.annotation.NonNull;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentTransaction;
-import android.support.v4.content.res.ResourcesCompat;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.LinearLayoutCompat;
@@ -20,7 +18,6 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.ViewTreeObserver;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageView;
@@ -28,12 +25,15 @@ import android.widget.LinearLayout;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
+import com.crashlytics.android.Crashlytics;
 import com.credolabs.justcredo.HorizontalListViewFragment;
 import com.credolabs.justcredo.R;
 import com.credolabs.justcredo.ReadReviewActivity;
 import com.credolabs.justcredo.ReviewActivity;
 import com.credolabs.justcredo.adapters.TextViewAdapter;
+import com.credolabs.justcredo.enums.PageTypes;
 import com.credolabs.justcredo.internet.ConnectionUtil;
+import com.credolabs.justcredo.model.DbConstants;
 import com.credolabs.justcredo.model.Review;
 import com.credolabs.justcredo.model.School;
 import com.credolabs.justcredo.model.User;
@@ -42,6 +42,7 @@ import com.credolabs.justcredo.newplace.BoardsFragment;
 import com.credolabs.justcredo.newplace.PlaceTypes;
 import com.credolabs.justcredo.newplace.TypePlaceFragment;
 import com.credolabs.justcredo.sliderlayout.ImageSlideAdapter;
+import com.credolabs.justcredo.utility.Constants;
 import com.credolabs.justcredo.utility.CustomToast;
 import com.credolabs.justcredo.utility.ExpandableHeightGridView;
 import com.credolabs.justcredo.utility.NearByPlaces;
@@ -55,16 +56,21 @@ import com.google.android.gms.location.places.ui.PlaceAutocomplete;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.ChildEventListener;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
+import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.Query;
+import com.google.firebase.firestore.WriteBatch;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Timer;
 import java.util.TimerTask;
 
@@ -74,17 +80,12 @@ import static android.app.Activity.RESULT_OK;
 public class SchoolHomeFragment extends Fragment {
 
     private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
+    private static final String TAG = "SchoolHomeFragment";
     private School model;
-    private String mParam2;
-
-    private EditText editTextName, editTextEmail, editTextWebsite, descriptionText, addressLine1, addressLine2, addressCity, addressState, addressCountry, mobileNumber;
+    private EditText editTextName, editTextEmail, editTextWebsite, descriptionText, addressLine1,
+            addressLine2, addressCity, addressState, addressCountry, mobileNumber;
     private static final int PLACE_AUTOCOMPLETE_REQUEST_CODE = 1;
-
     private LinearLayout school_edit_home;
-
-
-    private OnFragmentInteractionListener mListener;
     private LatLng latLng;
     private Button save_school_home;
     private ProgressDialog mProgressDialog;
@@ -93,15 +94,17 @@ public class SchoolHomeFragment extends Fragment {
     private FirebaseUser user;
     private LinearLayout fragment_container;
     private Button edit_school_home,cancel_school_home;
+    private Query query;
+    private Review review1,review2,review3;
+    private User user1,user2,user3;
 
     public SchoolHomeFragment() {
     }
 
-    public static SchoolHomeFragment newInstance(School param1, String param2) {
+    public static SchoolHomeFragment newInstance(School param1) {
         SchoolHomeFragment fragment = new SchoolHomeFragment();
         Bundle args = new Bundle();
         args.putSerializable(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
         fragment.setArguments(args);
         return fragment;
     }
@@ -111,25 +114,22 @@ public class SchoolHomeFragment extends Fragment {
         super.onCreate(savedInstanceState);
         if (getArguments() != null) {
             model = (School) getArguments().getSerializable(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
         }
-
         FragmentTransaction transaction = getChildFragmentManager().beginTransaction();
-        Fragment top_rated = HorizontalListViewFragment.newInstance(PlaceTypes.PageTypes.DETAIL_PAGE.getValue(),model.getType(),model);
+        Fragment top_rated = HorizontalListViewFragment.newInstance(PageTypes.DETAIL_PAGE.getValue(),model.getType(),model);
         transaction.add(R.id.fragment_container, top_rated );
         transaction.commit();
 
     }
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         final View view = inflater.inflate(R.layout.fragment_school_home, container, false);
-        ConnectionUtil.checkConnection(getActivity().findViewById(R.id.placeSnackBar));
+        if (getActivity()!=null)ConnectionUtil.checkConnection(getActivity().findViewById(R.id.placeSnackBar));
 
-        final ViewPager mPager = (ViewPager) view.findViewById(R.id.view_pager);
+        final ViewPager mPager = view.findViewById(R.id.view_pager);
         if (model.getImages()!=null) {
-            final ArrayList<String> list = new ArrayList<String>(model.getImages().values());
+            final ArrayList<String> list = new ArrayList<>(model.getImages().values());
             String address = Util.getAddress(model.getAddress());
             ZoomObject zoomObject = new ZoomObject();
             zoomObject.setImages(list);
@@ -138,15 +138,12 @@ public class SchoolHomeFragment extends Fragment {
             zoomObject.setLogo(list.get(0));
             mPager.setAdapter(new ImageSlideAdapter(getActivity(), list, zoomObject));
             final int[] currentPage = {0};
-            // Auto start of viewpager
             final Handler handler = new Handler();
-            final Runnable Update = new Runnable() {
-                public void run() {
-                    if (currentPage[0] == list.size()) {
-                        currentPage[0] = 0;
-                    }
-                    mPager.setCurrentItem(currentPage[0]++, true);
+            final Runnable Update = () -> {
+                if (currentPage[0] == list.size()) {
+                    currentPage[0] = 0;
                 }
+                mPager.setCurrentItem(currentPage[0]++, true);
             };
             Timer swipeTimer = new Timer();
             swipeTimer.schedule(new TimerTask() {
@@ -158,111 +155,95 @@ public class SchoolHomeFragment extends Fragment {
         }else {
             mPager.setVisibility(View.GONE);
         }
-        //TextView edit_button = (TextView) view.findViewById(R.id.edit_button);
-        fragment_container = (LinearLayout) view.findViewById(R.id.fragment_container);
-        school_edit_home = (LinearLayout) view.findViewById(R.id.school_edit_home);
-        LinearLayout edit_school_home_section = (LinearLayout) view.findViewById(R.id.edit_school_home_section);
-        cancel_school_home = (Button) view.findViewById(R.id.cancel_school_home);
-        edit_school_home = (Button) view.findViewById(R.id.edit_school_home);
+        fragment_container = view.findViewById(R.id.fragment_container);
+        school_edit_home = view.findViewById(R.id.school_edit_home);
+        LinearLayout edit_school_home_section = view.findViewById(R.id.edit_school_home_section);
+        cancel_school_home = view.findViewById(R.id.cancel_school_home);
+        edit_school_home = view.findViewById(R.id.edit_school_home);
         FirebaseAuth auth = FirebaseAuth.getInstance();
         user = auth.getCurrentUser();
-        editTextName    = (EditText) view.findViewById(R.id.EditTextName);
-        editTextEmail   = (EditText) view.findViewById(R.id.EditTextEmail);
-        editTextWebsite = (EditText) view.findViewById(R.id.EditTextWebsite);
-        descriptionText = (EditText) view.findViewById(R.id.description_text);
-        LinearLayout addressLayout = (LinearLayout) view.findViewById(R.id.addressLayout);
-        addressLine1    = (EditText) view.findViewById(R.id.addressLine1);
-        addressLine2    = (EditText) view.findViewById(R.id.addressLine2);
-        addressCity     = (EditText) view.findViewById(R.id.addressCity);
-        addressState    = (EditText) view.findViewById(R.id.addressState);
-        addressCountry  = (EditText) view.findViewById(R.id.addressCountry);
-        mobileNumber    = (EditText) view.findViewById(R.id.mobileNumber);
-        save_school_home = (Button) view.findViewById(R.id.save_school_home);
-        bookmark = (LinearLayout) view.findViewById(R.id.bookmark);
-        bookmarkImage = (ImageView) view.findViewById(R.id.bookmarkImage);
+        editTextName    = view.findViewById(R.id.EditTextName);
+        editTextEmail   = view.findViewById(R.id.EditTextEmail);
+        editTextWebsite = view.findViewById(R.id.EditTextWebsite);
+        descriptionText = view.findViewById(R.id.description_text);
+        addressLine1    = view.findViewById(R.id.addressLine1);
+        addressLine2    = view.findViewById(R.id.addressLine2);
+        addressCity     = view.findViewById(R.id.addressCity);
+        addressState    = view.findViewById(R.id.addressState);
+        addressCountry  = view.findViewById(R.id.addressCountry);
+        mobileNumber    = view.findViewById(R.id.mobileNumber);
+        save_school_home = view.findViewById(R.id.save_school_home);
+        bookmark = view.findViewById(R.id.bookmark);
+        bookmarkImage = view.findViewById(R.id.bookmarkImage);
 
         setBookmark();
 
         if (Util.checkSchoolAdmin(model)){
             edit_school_home_section.setVisibility(View.VISIBLE);
-            edit_school_home.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    fragment_container.setVisibility(View.GONE);
-                    school_edit_home.setVisibility(View.VISIBLE);
-                    edit_school_home.setVisibility(View.GONE);
-                    cancel_school_home.setVisibility(View.VISIBLE);
-                    buildEditSection(view);
+            edit_school_home.setOnClickListener(v -> {
+                fragment_container.setVisibility(View.GONE);
+                school_edit_home.setVisibility(View.VISIBLE);
+                edit_school_home.setVisibility(View.GONE);
+                cancel_school_home.setVisibility(View.VISIBLE);
+                buildEditSection(view);
 
 
-                }
             });
 
-            cancel_school_home.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    fragment_container.setVisibility(View.VISIBLE);
-                    school_edit_home.setVisibility(View.GONE);
-                    edit_school_home.setVisibility(View.VISIBLE);
-                    cancel_school_home.setVisibility(View.GONE);
-                }
+            cancel_school_home.setOnClickListener(v -> {
+                fragment_container.setVisibility(View.VISIBLE);
+                school_edit_home.setVisibility(View.GONE);
+                edit_school_home.setVisibility(View.VISIBLE);
+                cancel_school_home.setVisibility(View.GONE);
             });
         }
-        final LinearLayout layoutReviews = (LinearLayout) view.findViewById(R.id.layout_reviews);
-        layoutReviews.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getActivity(),ReadReviewActivity.class);
-                intent.putExtra("key",model.getId());
-                intent.putExtra("type","schools");
-                startActivity(intent);
-            }
+        final LinearLayout layoutReviews = view.findViewById(R.id.layout_reviews);
+        layoutReviews.setOnClickListener(v -> {
+            Intent intent = new Intent(getActivity(),ReadReviewActivity.class);
+            intent.putExtra(School.ID,model.getId());
+            intent.putExtra(School.NAME,model.getName());
+            startActivity(intent);
         });
 
-        TextView reviewBtn = (TextView) view.findViewById(R.id.btn_review);
-        reviewBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getActivity(),ReviewActivity.class);
-                intent.putExtra("name",model.getName());
-                intent.putExtra("type","schools");
-                intent.putExtra("id",model.getId());
-                intent.putExtra("addressCity",model.getAddress().get("addressCity"));
-                intent.putExtra("addressState",model.getAddress().get("addressState"));
-                startActivity(intent);
+        TextView reviewBtn = view.findViewById(R.id.btn_review);
+        reviewBtn.setOnClickListener(v -> {
+            Intent intent = new Intent(getActivity(),ReviewActivity.class);
+            intent.putExtra(School.NAME,model.getName());
+            intent.putExtra(School.TYPE,School.SCHOOL_DATABASE);
+            intent.putExtra(School.ID,model.getId());
+            intent.putExtra(School.ADDRESS_CITY,model.getAddress().get(School.ADDRESS_CITY));
+            intent.putExtra(School.ADDRESS_STATE,model.getAddress().get(School.ADDRESS_STATE));
+            startActivity(intent);
 
-            }
         });
 
-        // for school header details
-        TextView schoolName = (TextView) view.findViewById(R.id.school_name);
+        TextView schoolName = view.findViewById(R.id.school_name); // for school header details
         if (model.getStatus()!=null && model.getStatus().equals(School.StatusType.VERIFIED.getValue())){
             schoolName.setCompoundDrawablesWithIntrinsicBounds(0,0,R.drawable.ic_verified,0);
         }
         schoolName.setText(model.getName());
-        TextView schoolAddress = (TextView) view.findViewById(R.id.school_address);
+        TextView schoolAddress = view.findViewById(R.id.school_address);
         schoolAddress.setText(Util.getAddress(model.getAddress()));
-        TextView schoolReviewNo = (TextView) view.findViewById(R.id.school_review_no);
+        TextView schoolReviewNo = view.findViewById(R.id.school_review_no);
         if (model.getNoOfReview()!=null){
             schoolReviewNo.setText(String.valueOf(model.getNoOfReview()));
         }
 
-        TextView school_bookmark_no = (TextView) view.findViewById(R.id.school_bookmark_no);
+        TextView school_bookmark_no = view.findViewById(R.id.school_bookmark_no);
         if (model.getNoOfBookmarks()!=null){
             school_bookmark_no.setText(String.valueOf(model.getNoOfBookmarks()));
         }
 
-        TextView distance = (TextView) view.findViewById(R.id.distance);
+        TextView distance = view.findViewById(R.id.distance);
         if (model.getLatitude()!=0 && model.getLongitude()!=0){
             distance.setText(getResources().getString(R.string.km,NearByPlaces.distance(getActivity(),model.getLatitude(),model.getLongitude())));
         }
 
-        // for school description
-        TextView schoolDescription = (TextView) view.findViewById(R.id.description);
+        TextView schoolDescription = view.findViewById(R.id.description); // for school description
         schoolDescription.setText(model.getDescription());
-        LinearLayout schoolCategory = (LinearLayout) view.findViewById(R.id.category_layout);
-        LinearLayout board_layout = (LinearLayout) view.findViewById(R.id.board_layout);
-        LinearLayout schoolGender = (LinearLayout) view.findViewById(R.id.gender);
+        LinearLayout schoolCategory = view.findViewById(R.id.category_layout);
+        LinearLayout board_layout = view.findViewById(R.id.board_layout);
+        LinearLayout schoolGender = view.findViewById(R.id.gender);
         if (model.getCategories()!=null){
             for (String value : model.getCategories().values()) {
                 TextView htext =new TextView(getActivity());
@@ -301,180 +282,96 @@ public class SchoolHomeFragment extends Fragment {
         }
 
 
-        // rating and review section
-        TextView schoolRating = (TextView) view.findViewById(R.id.rating);
-        TextView noOfRatings = (TextView) view.findViewById(R.id.noOfRatings);
-        LinearLayoutCompat layoutRating = (LinearLayoutCompat) view.findViewById(R.id.layout_rating);
+        TextView schoolRating = view.findViewById(R.id.rating);   // rating and review section
+        TextView noOfRatings = view.findViewById(R.id.noOfRatings);
+        LinearLayoutCompat layoutRating = view.findViewById(R.id.layout_rating);
         if (model.getNoOfRating() ==null || model.getNoOfRating()==0){
             layoutRating.setVisibility(View.GONE);
         }else {
             schoolRating.setText(String.valueOf(model.getRating()));
             noOfRatings.setText(String.valueOf(model.getNoOfRating()));
         }
-        final ImageView profile1 = (ImageView) view.findViewById(R.id.profile1);
-        final ImageView profile2 = (ImageView) view.findViewById(R.id.profile2);
-        final ImageView profile3 = (ImageView) view.findViewById(R.id.profile3);
-        String nameFirstUser = " ";
-        //final LinearLayout layoutReviews = (LinearLayout) view.findViewById(R.id.layout_reviews);
-        final LinearLayout layoutReviewsNo = (LinearLayout) view.findViewById(R.id.layout_reviews_no);
+        final ImageView profile1 = view.findViewById(R.id.profile1);
+        final ImageView profile2 = view.findViewById(R.id.profile2);
+        final ImageView profile3 = view.findViewById(R.id.profile3);
+        final LinearLayout layoutReviewsNo = view.findViewById(R.id.layout_reviews_no);
 
-        TextView noOfReviewsMinusOne = (TextView) view.findViewById(R.id.noOfReviewsMinusOne);
-        if (model.getNoOfReview()!=null && model.getNoOfReview()!=0){
-            noOfReviewsMinusOne.setText(String.valueOf(model.getNoOfReview()-1));
-            TextView allReview = (TextView) view.findViewById(R.id.all_review);
-            allReview.setText("Read All Reviews ("+model.getNoOfReview()+")");
-        }
-
-        DatabaseReference mDatabaseReference = FirebaseDatabase.getInstance().getReference().child("reviews");
-        final TextView reviewerName1 = (TextView) view.findViewById(R.id.reviewer_name1);
-
-        final ArrayList<Review> reviews= new ArrayList<>() ;
-        mDatabaseReference.orderByChild("schoolID").equalTo(model.getId()).addChildEventListener(new ChildEventListener() {
-            @Override
-            public void onChildAdded(DataSnapshot dataSnapshot, String s) {
-                Review review = dataSnapshot.getValue(Review.class);
-                reviews.add(review);
-                switch (reviews.size()){
-                    case 1:
+        final TextView reviewerName1 = view.findViewById(R.id.reviewer_name1);
+        FirebaseFirestore db = FirebaseFirestore.getInstance();
+        db.collection(Review.DB_REVIEWS_REF)
+                .whereEqualTo(Review.SCHOOL_ID,model.getId()).get().addOnCompleteListener(task -> {
+                if (task.isSuccessful() && task.getResult()!=null){
+                    List<DocumentSnapshot> documents = task.getResult().getDocuments();
+                    if (documents.size()>0){
                         layoutReviews.setVisibility(View.VISIBLE);
                         layoutReviewsNo.setVisibility(View.GONE);
-                        profile1.setVisibility(View.VISIBLE);
-                        DatabaseReference userReference = FirebaseDatabase.getInstance().getReference().child("users");
-                        if (review != null) {
-                            userReference.orderByKey().equalTo(review.getUserID()).addChildEventListener(new ChildEventListener() {
-                                @Override
-                                public void onChildAdded(DataSnapshot dataSnapshot, String s) {
-                                    User user = dataSnapshot.getValue(User.class);
-                                    //profile1.setImageUrl(user.getProfilePic(),imgLoader);
-                                    if (user != null) {
-                                        Util.loadCircularImageWithGlide(getActivity(),user.getProfilePic(),profile1);
-                                    }
-                                    reviewerName1.setText(user.getName());
+                    }
+                    switch (documents.size()){
+                        case 1:
+                            review1 = documents.get(0).toObject(Review.class);
+                            query = db.collection(User.DB_REF).whereEqualTo(User.UID,review1.getUserID());
+                            break;
+                        case 2:
+                            review1 = documents.get(0).toObject(Review.class);
+                            review2 = documents.get(1).toObject(Review.class);
+                            query = db.collection(User.DB_REF).whereEqualTo(User.UID,review1.getUserID())
+                                                              .whereEqualTo(User.UID,review2.getUserID());
+                            break;
+                        default:
+                            review1 = documents.get(0).toObject(Review.class);
+                            review2 = documents.get(1).toObject(Review.class);
+                            review3 = documents.get(2).toObject(Review.class);
+                            query = db.collection(User.DB_REF).whereEqualTo(User.UID,review1.getUserID())
+                                                              .whereEqualTo(User.UID,review2.getUserID())
+                                                              .whereEqualTo(User.UID,review3.getUserID());
+                            break;
+                    }
+                    query.get().addOnCompleteListener(userTask -> {
+                        if (userTask.isSuccessful() && userTask.getResult() != null) {
+                            List<DocumentSnapshot> userDocument = userTask.getResult().getDocuments();
+                            for (int i=0;i<userDocument.size();i++) {
+                                switch (i){
+                                    case 0:
+                                        user1 = userDocument.get(i).toObject(User.class);
+                                        if (user1 != null) {
+                                            profile1.setVisibility(View.VISIBLE);
+                                            Util.loadCircularImageWithGlide(getActivity(),user1.getProfilePic(),profile1);
+                                            reviewerName1.setText(String.format(getString(R.string.review_no_msg_for_1), user1.getName()));
+                                        }
+                                        break;
+                                    case 1:
+                                        user2 = userDocument.get(i).toObject(User.class);
+                                        if (user2 != null) {
+                                            if (user1.getUid().equals(user2.getUid())){
+                                                reviewerName1.setText(String.format(getString(R.string.review_no_msg_for_1), user1.getName()));
+                                            }else {
+                                                reviewerName1.setText(String.format(getString(R.string.review_no_msg_for_2), user1.getName()));
+                                                profile2.setVisibility(View.VISIBLE);
+                                                Util.loadCircularImageWithGlide(getActivity(),user2.getProfilePic(),profile2);
+                                            }
+                                        }
+                                        break;
+                                    case 2:
+                                        user3 = userDocument.get(i).toObject(User.class);
+                                        if (user3 != null) {
+                                            if (user3.getUid().equals(user1.getUid()) || user3.getUid().equals(user2.getUid())){
+                                                reviewerName1.setText(String.format(getString(R.string.review_no_msg_for_2), user1.getName()));
+                                            }else {
+                                                profile3.setVisibility(View.VISIBLE);
+                                                Util.loadCircularImageWithGlide(getActivity(),user3.getProfilePic(),profile3);
+                                                reviewerName1.setText(String.format(getString(R.string.review_no_msg_for_greater), user1.getName(), model.getNoOfReview() -1 ));
+                                            }
+                                        }
+                                        break;
                                 }
-
-                                @Override
-                                public void onChildChanged(DataSnapshot dataSnapshot, String s) {
-
-                                }
-
-                                @Override
-                                public void onChildRemoved(DataSnapshot dataSnapshot) {
-
-                                }
-
-                                @Override
-                                public void onChildMoved(DataSnapshot dataSnapshot, String s) {
-
-                                }
-
-                                @Override
-                                public void onCancelled(DatabaseError databaseError) {
-
-                                }
-                            });
+                            }
                         }
-                        break;
-                    case 2:
-                        profile2.setVisibility(View.VISIBLE);
-                        DatabaseReference userReference2 = FirebaseDatabase.getInstance().getReference().child("users");
-                        if (review != null) {
-                            userReference2.orderByKey().equalTo(review.getUserID()).addChildEventListener(new ChildEventListener() {
-                                @Override
-                                public void onChildAdded(DataSnapshot dataSnapshot, String s) {
-                                    User user = dataSnapshot.getValue(User.class);
-                                    //profile2.setImageUrl(user.getProfilePic(),imgLoader);
-                                    if (user != null) {
-                                        Util.loadCircularImageWithGlide(getActivity(),user.getProfilePic(),profile2);
-                                    }
-
-                                }
-
-                                @Override
-                                public void onChildChanged(DataSnapshot dataSnapshot, String s) {
-
-                                }
-
-                                @Override
-                                public void onChildRemoved(DataSnapshot dataSnapshot) {
-
-                                }
-
-                                @Override
-                                public void onChildMoved(DataSnapshot dataSnapshot, String s) {
-
-                                }
-
-                                @Override
-                                public void onCancelled(DatabaseError databaseError) {
-
-                                }
-                            });
-                        }
-                        break;
-                    case 3:
-                        profile3.setVisibility(View.VISIBLE);
-                        DatabaseReference userReference3 = FirebaseDatabase.getInstance().getReference().child("users");
-                        if (review != null) {
-                            userReference3.orderByKey().equalTo(review.getUserID()).addChildEventListener(new ChildEventListener() {
-                                @Override
-                                public void onChildAdded(DataSnapshot dataSnapshot, String s) {
-                                    User user = dataSnapshot.getValue(User.class);
-                                    //profile3.setImageUrl(user.getProfilePic(),imgLoader);
-                                    if (user != null) {
-                                        Util.loadCircularImageWithGlide(getActivity(),user.getProfilePic(),profile3);
-                                    }
-
-                                }
-
-                                @Override
-                                public void onChildChanged(DataSnapshot dataSnapshot, String s) {
-
-                                }
-
-                                @Override
-                                public void onChildRemoved(DataSnapshot dataSnapshot) {
-
-                                }
-
-                                @Override
-                                public void onChildMoved(DataSnapshot dataSnapshot, String s) {
-
-                                }
-
-                                @Override
-                                public void onCancelled(DatabaseError databaseError) {
-
-                                }
-                            });
-                        }
-                        break;
+                    });
                 }
-
-            }
-
-            @Override
-            public void onChildChanged(DataSnapshot dataSnapshot, String s) {
-
-            }
-
-            @Override
-            public void onChildRemoved(DataSnapshot dataSnapshot) {
-
-            }
-
-            @Override
-            public void onChildMoved(DataSnapshot dataSnapshot, String s) {
-
-            }
-
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-
-            }
         });
 
-        LinearLayout layout_classes = (LinearLayout) view.findViewById(R.id.layout_classes);
-        ExpandableHeightGridView expandableHeightGridView = (ExpandableHeightGridView) view.findViewById(R.id.checked);
+        LinearLayout layout_classes = view.findViewById(R.id.layout_classes);
+        ExpandableHeightGridView expandableHeightGridView = view.findViewById(R.id.checked);
         if (model.getClasses()!=null){
             ArrayList<String> classList = new ArrayList<>(model.getClasses().values());
             Collections.sort(classList);
@@ -484,168 +381,119 @@ public class SchoolHomeFragment extends Fragment {
         }
 
 
-        // Section for editing place by admin
-        ImageView btn_edit_classes = (ImageView) view.findViewById(R.id.btn_edit_classes);
-        ImageView btn_edit_type = (ImageView) view.findViewById(R.id.btn_edit_type);
+        ImageView btn_edit_classes = view.findViewById(R.id.btn_edit_classes); // Section for editing place by admin
+        ImageView btn_edit_type = view.findViewById(R.id.btn_edit_type);
         if (Util.checkSchoolAdmin(model)) {
             final BoardsFragment boardsFragment = BoardsFragment.newInstance(PlaceTypes.Action.EDIT_BACKUP.getValue(), model);
-            final LinearLayout edit_boards_section = (LinearLayout) view.findViewById(R.id.edit_boards_section);
-            final LinearLayout edit_type_section = (LinearLayout) view.findViewById(R.id.edit_type_section);
-            final LinearLayout content = (LinearLayout) view.findViewById(R.id.content);
-            Button save_boards = (Button) view.findViewById(R.id.save_boards);
-            Button cancel_boards = (Button) view.findViewById(R.id.cancel_boards);
-            final LinearLayout fragContainer = (LinearLayout) view.findViewById(R.id.container);
+            final LinearLayout edit_boards_section = view.findViewById(R.id.edit_boards_section);
+            final LinearLayout edit_type_section = view.findViewById(R.id.edit_type_section);
+            final LinearLayout content = view.findViewById(R.id.content);
+            Button save_boards = view.findViewById(R.id.save_boards);
+            Button cancel_boards = view.findViewById(R.id.cancel_boards);
+            final LinearLayout fragContainer = view.findViewById(R.id.container);
             fragContainer.setVisibility(View.GONE);
-            final ProgressBar progress = (ProgressBar) view.findViewById(R.id.progress);
-            btn_edit_classes.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    progress.setVisibility(View.VISIBLE);
-                    FragmentTransaction transaction = getChildFragmentManager().beginTransaction();
-                    transaction.replace(R.id.container, boardsFragment);
-                    transaction.addToBackStack(null);
-                    transaction.commit();
-                    edit_boards_section.setVisibility(View.VISIBLE);
-                    edit_type_section.setVisibility(View.GONE);
-                    content.setVisibility(View.GONE);
-                    fragContainer.setVisibility(View.VISIBLE);
-                    progress.setVisibility(View.GONE);
-                }
+            final ProgressBar progress = view.findViewById(R.id.progress);
+            btn_edit_classes.setOnClickListener(v -> {
+                progress.setVisibility(View.VISIBLE);
+                FragmentTransaction transaction = getChildFragmentManager().beginTransaction();
+                transaction.replace(R.id.container, boardsFragment);
+                transaction.addToBackStack(null);
+                transaction.commit();
+                edit_boards_section.setVisibility(View.VISIBLE);
+                edit_type_section.setVisibility(View.GONE);
+                content.setVisibility(View.GONE);
+                fragContainer.setVisibility(View.VISIBLE);
+                progress.setVisibility(View.GONE);
             });
 
-            cancel_boards.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    edit_boards_section.setVisibility(View.GONE);
-                    fragContainer.setVisibility(View.GONE);
-                    content.setVisibility(View.VISIBLE);
-                }
+            cancel_boards.setOnClickListener(v -> {
+                edit_boards_section.setVisibility(View.GONE);
+                fragContainer.setVisibility(View.GONE);
+                content.setVisibility(View.VISIBLE);
             });
 
-            save_boards.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(getActivity());
-                    alertDialogBuilder.setMessage("Do you want to update?");
-                    alertDialogBuilder.setPositiveButton("Yes",
-                            new DialogInterface.OnClickListener() {
-                                @Override
-                                public void onClick(DialogInterface arg0, int arg1) {
-                                    progress.setVisibility(View.VISIBLE);
-                                    HashMap<String, HashMap<String, String>> map = boardsFragment.getFragmentState();
-                                    DatabaseReference mDatabaseReference = FirebaseDatabase.getInstance().getReference().child(School.SCHOOL_DATABASE);
-
-                                    mDatabaseReference.child(model.getId()).child(School.BOARDS).setValue(map.get(School.BOARDS));
-                                    mDatabaseReference.child(model.getId()).child(School.CLASSES).setValue(map.get(School.CLASSES));
-                                    fragContainer.setVisibility(View.GONE);
-                                    content.setVisibility(View.VISIBLE);
-                                    edit_boards_section.setVisibility(View.GONE);
-                                    progress.setVisibility(View.GONE);
-                                    AlertDialog.Builder builder = new AlertDialog.Builder(
-                                            getActivity());
-                                    builder.setCancelable(true);
-                                    builder.setMessage("Congrats, place updated successfully ! You can write blog and upload photos so that users" +
-                                            " can know about this.");
-                                    builder.setPositiveButton("Ok",
-                                            new DialogInterface.OnClickListener() {
-                                                @Override
-                                                public void onClick(DialogInterface dialog,
-                                                                    int which) {
-                                                    dialog.dismiss();
-                                                }
-                                            });
-                                    AlertDialog alert = builder.create();
-                                    alert.show();
-                                }
+            save_boards.setOnClickListener(v -> {
+                AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(getActivity());
+                alertDialogBuilder.setMessage(R.string.update_confirmation_msg);
+                alertDialogBuilder.setPositiveButton(R.string.yes,
+                        (arg0, arg1) -> {
+                            progress.setVisibility(View.VISIBLE);
+                            HashMap<String, HashMap<String, String>> map = boardsFragment.getFragmentState();
+                            WriteBatch batch = db.batch();
+                            DocumentReference documentReference = db.collection(School.SCHOOL_DATABASE).document(model.getId());
+                            batch.update(documentReference, School.BOARDS, map.get(School.BOARDS));
+                            batch.update(documentReference, School.CLASSES, map.get(School.CLASSES));
+                            batch.commit().addOnCompleteListener(task -> {
+                                fragContainer.setVisibility(View.GONE);
+                                content.setVisibility(View.VISIBLE);
+                                edit_boards_section.setVisibility(View.GONE);
+                                progress.setVisibility(View.GONE);
+                                AlertDialog.Builder builder = new AlertDialog.Builder(
+                                        getActivity());
+                                builder.setCancelable(true);
+                                builder.setMessage(R.string.place_updated_success_msg);
+                                builder.setPositiveButton(R.string.ok,
+                                        (dialog, which) -> dialog.dismiss());
+                                AlertDialog alert = builder.create();
+                                alert.show();
                             });
-
-                    alertDialogBuilder.setNegativeButton("No", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            dialog.dismiss();
-                        }
-                    });
-
-                    alertDialogBuilder.create();
-                    alertDialogBuilder.show();
-                }
+                        });
+                alertDialogBuilder.setNegativeButton(R.string.no, (dialog, which) -> dialog.dismiss());
+                alertDialogBuilder.create();
+                alertDialogBuilder.show();
             });
 
             final TypePlaceFragment typePlaceFragment = TypePlaceFragment.newInstance(PlaceTypes.Action.EDIT_BACKUP.getValue(), model);
-            Button save_type = (Button) view.findViewById(R.id.save_type);
-            Button cancel_type = (Button) view.findViewById(R.id.cancel_type);
-            btn_edit_type.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    progress.setVisibility(View.VISIBLE);
-                    FragmentTransaction transaction = getChildFragmentManager().beginTransaction();
-                    transaction.replace(R.id.container, typePlaceFragment);
-                    transaction.addToBackStack(null);
-                    transaction.commit();
-                    edit_type_section.setVisibility(View.VISIBLE);
-                    edit_boards_section.setVisibility(View.GONE);
-                    content.setVisibility(View.GONE);
-                    fragContainer.setVisibility(View.VISIBLE);
-                    progress.setVisibility(View.GONE);
-                }
+            Button save_type = view.findViewById(R.id.save_type);
+            Button cancel_type = view.findViewById(R.id.cancel_type);
+            btn_edit_type.setOnClickListener(v -> {
+                progress.setVisibility(View.VISIBLE);
+                FragmentTransaction transaction = getChildFragmentManager().beginTransaction();
+                transaction.replace(R.id.container, typePlaceFragment);
+                transaction.addToBackStack(null);
+                transaction.commit();
+                edit_type_section.setVisibility(View.VISIBLE);
+                edit_boards_section.setVisibility(View.GONE);
+                content.setVisibility(View.GONE);
+                fragContainer.setVisibility(View.VISIBLE);
+                progress.setVisibility(View.GONE);
             });
 
-            cancel_type.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    edit_type_section.setVisibility(View.GONE);
-                    fragContainer.setVisibility(View.GONE);
-                    content.setVisibility(View.VISIBLE);
-                }
+            cancel_type.setOnClickListener(v -> {
+                edit_type_section.setVisibility(View.GONE);
+                fragContainer.setVisibility(View.GONE);
+                content.setVisibility(View.VISIBLE);
             });
 
-            save_type.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(getActivity());
-                    alertDialogBuilder.setMessage("Do you want to update?");
-                    alertDialogBuilder.setPositiveButton("Yes",
-                            new DialogInterface.OnClickListener() {
-                                @Override
-                                public void onClick(DialogInterface arg0, int arg1) {
-                                    progress.setVisibility(View.VISIBLE);
-                                    HashMap<String, HashMap<String, String>> map = typePlaceFragment.getFragmentState();
-                                    DatabaseReference mDatabaseReference = FirebaseDatabase.getInstance().getReference().child(School.SCHOOL_DATABASE);
-
-                                    mDatabaseReference.child(model.getId()).child(School.CATEGORIES).setValue(map.get(School.CATEGORIES));
-                                    mDatabaseReference.child(model.getId()).child(School.GENDER).setValue(map.get(School.GENDER));
-                                    fragContainer.setVisibility(View.GONE);
-                                    content.setVisibility(View.VISIBLE);
-                                    edit_type_section.setVisibility(View.GONE);
-                                    progress.setVisibility(View.GONE);
-                                    AlertDialog.Builder builder = new AlertDialog.Builder(
-                                            getActivity());
-                                    builder.setCancelable(true);
-                                    builder.setMessage("Congrats, place updated successfully ! You can write blog and upload photos so that users" +
-                                            " can know about this.");
-                                    builder.setPositiveButton("Ok",
-                                            new DialogInterface.OnClickListener() {
-                                                @Override
-                                                public void onClick(DialogInterface dialog,
-                                                                    int which) {
-                                                    dialog.dismiss();
-                                                }
-                                            });
-                                    AlertDialog alert = builder.create();
-                                    alert.show();
-                                }
+            save_type.setOnClickListener(v -> {
+                AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(getActivity());
+                alertDialogBuilder.setMessage(R.string.update_confirmation_msg);
+                alertDialogBuilder.setPositiveButton(R.string.yes,
+                        (arg0, arg1) -> {
+                            progress.setVisibility(View.VISIBLE);
+                            HashMap<String, HashMap<String, String>> map = typePlaceFragment.getFragmentState();
+                            WriteBatch batch = db.batch();
+                            DocumentReference documentReference = db.collection(School.SCHOOL_DATABASE).document(model.getId());
+                            batch.update(documentReference, School.CATEGORIES, map.get(School.CATEGORIES));
+                            batch.update(documentReference, School.GENDER, map.get(School.GENDER));
+                            batch.commit().addOnCompleteListener(task -> {
+                                fragContainer.setVisibility(View.GONE);
+                                content.setVisibility(View.VISIBLE);
+                                edit_type_section.setVisibility(View.GONE);
+                                progress.setVisibility(View.GONE);
+                                AlertDialog.Builder builder = new AlertDialog.Builder(
+                                        getActivity());
+                                builder.setCancelable(true);
+                                builder.setMessage(R.string.place_updated_success_msg);
+                                builder.setPositiveButton(R.string.ok,
+                                        (dialog, which) -> dialog.dismiss());
+                                AlertDialog alert = builder.create();
+                                alert.show();
                             });
-
-                    alertDialogBuilder.setNegativeButton("No", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            dialog.dismiss();
-                        }
-                    });
-
-                    alertDialogBuilder.create();
-                    alertDialogBuilder.show();
-                }
+                        });
+                alertDialogBuilder.setNegativeButton(R.string.no, (dialog, which) -> dialog.dismiss());
+                alertDialogBuilder.create();
+                alertDialogBuilder.show();
             });
         }else {
             btn_edit_classes.setVisibility(View.GONE);
@@ -656,33 +504,15 @@ public class SchoolHomeFragment extends Fragment {
     }
 
     private void setBookmark() {
-
-        final DatabaseReference mBookmarkReference = FirebaseDatabase.getInstance().getReference().child("bookmarks");
-
-        mBookmarkReference.child(user.getUid()).addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                if (dataSnapshot.hasChild(model.getId())){
-                    bookmarkImage.setImageResource(R.drawable.ic_bookmark_green_24dp);
-                }else{
-                    bookmarkImage.setImageResource(R.drawable.ic_bookmark_secondary);
-                }
-            }
-
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-
+        FirebaseFirestore.getInstance().collection(DbConstants.DB_REF_BOOKMARK).document(user.getUid())
+                .get().addOnCompleteListener(task -> {
+            if (task.isSuccessful() && task.getResult()!=null && task.getResult().contains(model.getId())){
+                bookmarkImage.setImageResource(R.drawable.ic_bookmark_green_24dp);
+            }else {
+                bookmarkImage.setImageResource(R.drawable.ic_bookmark_secondary);
             }
         });
-
-        bookmark.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                School.onBookmark(model,user,getActivity(),bookmarkImage);
-
-            }
-        });
-
+        bookmark.setOnClickListener(v -> School.onBookmark(model,user,getActivity(),bookmarkImage));
     }
 
 
@@ -693,42 +523,45 @@ public class SchoolHomeFragment extends Fragment {
             switch (requestCode) {
                 case PLACE_AUTOCOMPLETE_REQUEST_CODE:
                     if (resultCode == RESULT_OK) {
-                        Place place = PlaceAutocomplete.getPlace(getActivity(), imageReturnedIntent);
-                        String address = String.valueOf(place.getAddress());
-                        latLng = place.getLatLng();
-                        String[] str= address.split(",");
-                        int j = 1;
+                        if (getActivity()!=null) {
+                            Place place = PlaceAutocomplete.getPlace(getActivity(), imageReturnedIntent);
+                            String address = String.valueOf(place.getAddress());
+                            latLng = place.getLatLng();
+                            String[] str = address.split(",");
+                            int j = 1;
 
-                        addressLine1.setText("");
-                        addressLine2.setText("");
-                        addressCity.setText("");
-                        addressState.setText("");
-                        addressCountry.setText("");
+                            addressLine1.setText("");
+                            addressLine2.setText("");
+                            addressCity.setText("");
+                            addressState.setText("");
+                            addressCountry.setText("");
 
-                        for(int i=str.length-1; i >=0 ; i--){
-                            if(j==1){
-                                addressCountry.setText(str[i]);
-                            }else if (j==2){
-                                addressState.setText(str[i]);
-                            }else if (j==3){
-                                addressCity.setText(str[i]);
-                            }else if (j==4){
-                                addressLine2.setText(str[i]);
-                            }else if (j==5){
-                                addressLine1.setText(str[i]);
-                            }else {
-                                addressLine1.setText(addressLine1.getText().toString().concat(" "+str[i]));
+                            for (int i = str.length - 1; i >= 0; i--) {
+                                if (j == 1) {
+                                    addressCountry.setText(str[i]);
+                                } else if (j == 2) {
+                                    addressState.setText(str[i]);
+                                } else if (j == 3) {
+                                    addressCity.setText(str[i]);
+                                } else if (j == 4) {
+                                    addressLine2.setText(str[i]);
+                                } else if (j == 5) {
+                                    addressLine1.setText(str[i]);
+                                } else {
+                                    addressLine1.setText(addressLine1.getText().toString().concat(" " + str[i]));
+                                }
+                                j++;
                             }
-                            j++;
                         }
-
-                        //addressLayout.setVisibility(View.VISIBLE);
                     } else if (resultCode == PlaceAutocomplete.RESULT_ERROR) {
-                        Status status = PlaceAutocomplete.getStatus(getActivity(), imageReturnedIntent);
-                        Log.i("TAG", status.getStatusMessage());
+                        if (getActivity()!=null) {
+                            Status status = PlaceAutocomplete.getStatus(getActivity(), imageReturnedIntent);
+                            Log.i(TAG, status.getStatusMessage());
+                        }
                     } else if (resultCode == RESULT_CANCELED) {
-                        // The user canceled the operation.
+                        Log.i(TAG, "user cancelled the operation");
                     }
+
                     break;
             }
         }
@@ -740,48 +573,41 @@ public class SchoolHomeFragment extends Fragment {
         editTextEmail.setText(model.getMail());
         editTextWebsite.setText(model.getWebsite());
         descriptionText.setText(model.getDescription());
-        addressLine1.setText(model.getAddress().get("addressLine1"));
-        addressLine2.setText(model.getAddress().get("addressLine2"));
-        addressCity.setText(model.getAddress().get("addressCity"));
-        addressState.setText(model.getAddress().get("addressState"));
-        addressCountry.setText(model.getAddress().get("addressCountry"));
+        addressLine1.setText(model.getAddress().get(School.ADDRESSLINE1));
+        addressLine2.setText(model.getAddress().get(School.ADDRESSLINE2));
+        addressCity.setText(model.getAddress().get(School.ADDRESS_CITY));
+        addressState.setText(model.getAddress().get(School.ADDRESS_STATE));
+        addressCountry.setText(model.getAddress().get(School.ADDRESS_COUNTRY));
         mobileNumber.setText(model.getMobileNumber());
 
 
-        EditText adressText = (EditText) view.findViewById(R.id.adressText);
+        EditText adressText = view.findViewById(R.id.adressText);
         final Activity activity = getActivity();
-        adressText.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                try {
-                    AutocompleteFilter typeFilter = new AutocompleteFilter.Builder()
-                            .setCountry("IN")
-                            .setTypeFilter(AutocompleteFilter.TYPE_FILTER_GEOCODE)
-                            .build();
-                    Intent intent =
-                            new PlaceAutocomplete.IntentBuilder(PlaceAutocomplete.MODE_FULLSCREEN)
-                                    .setFilter(typeFilter)
-                                    .build(activity);
+        adressText.setOnClickListener(v -> {
+            try {
+                AutocompleteFilter typeFilter = new AutocompleteFilter.Builder()
+                        .setCountry(Constants.COUNTRY_INDIA)
+                        .setTypeFilter(AutocompleteFilter.TYPE_FILTER_GEOCODE)
+                        .build();
+                if (activity != null) {
+                    Intent intent = new PlaceAutocomplete.IntentBuilder(PlaceAutocomplete.MODE_FULLSCREEN)
+                            .setFilter(typeFilter)
+                            .build(activity);
                     startActivityForResult(intent, PLACE_AUTOCOMPLETE_REQUEST_CODE);
-                } catch (GooglePlayServicesRepairableException e) {
-                } catch (GooglePlayServicesNotAvailableException e) {
+
                 }
+            } catch (GooglePlayServicesRepairableException | GooglePlayServicesNotAvailableException e) {
+                Crashlytics.log(Log.INFO,TAG, e.getMessage());
             }
         });
 
         mProgressDialog = new ProgressDialog(getActivity());
-
-        save_school_home.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                boolean validationPassed = validate();
-                if (validationPassed) {
-                    startPosting();
-                }
+        save_school_home.setOnClickListener(v -> {
+            boolean validationPassed = validate();
+            if (validationPassed) {
+                startPosting();
             }
         });
-
-
     }
 
     private boolean validate() {
@@ -789,7 +615,7 @@ public class SchoolHomeFragment extends Fragment {
         if (TextUtils.isEmpty(name)){
             editTextName.requestFocus();
             new CustomToast().Show_Toast(getActivity(),
-                    "Please provide a name to this place!");
+                    getString(R.string.provide_placce_name));
             return false;
         }
 
@@ -797,14 +623,14 @@ public class SchoolHomeFragment extends Fragment {
         if (TextUtils.isEmpty(description)){
             descriptionText.requestFocus();
             new CustomToast().Show_Toast(getActivity(),
-                    "Please provide description to this place!");
+                    getString(R.string.provide_place_desc));
             return false;
         }
 
         if (description.length()< 100){
             descriptionText.requestFocus();
             new CustomToast().Show_Toast(getActivity(),
-                    "Description should be alteast 100 characters!");
+                    getString(R.string.desc_100_msg));
             return false;
         }
 
@@ -815,25 +641,25 @@ public class SchoolHomeFragment extends Fragment {
         if (TextUtils.isEmpty(city)){
             addressCity.requestFocus();
             new CustomToast().Show_Toast(getActivity(),
-                    "Please provide city in the address section!");
+                    getString(R.string.provide_city));
             return false;
         }
         if (TextUtils.isEmpty(state)){
             addressState.requestFocus();
             new CustomToast().Show_Toast(getActivity(),
-                    "Please provide state in the address section!");
+                    getString(R.string.provide_state));
             return false;
         }
         if (TextUtils.isEmpty(country)){
             addressCountry.requestFocus();
             new CustomToast().Show_Toast(getActivity(),
-                    "Please provide country in the address section!");
+                    getString(R.string.provide_country));
             return false;
         }
         if (TextUtils.isEmpty(mobile)){
             mobileNumber.requestFocus();
             new CustomToast().Show_Toast(getActivity(),
-                    "Please provide mobile number in the address section!");
+                    getString(R.string.provide_mobile));
             return false;
         }
 
@@ -857,110 +683,87 @@ public class SchoolHomeFragment extends Fragment {
         String mobile = mobileNumber.getText().toString().trim();
 
 
-        mProgressDialog.setMessage("Submitting Your Place");
+        mProgressDialog.setMessage(getString(R.string.submitting_your_place));
         mProgressDialog.setProgressStyle(ProgressDialog.STYLE_SPINNER);
         mProgressDialog.setIndeterminate(true);
         mProgressDialog.setCancelable(false);
         mProgressDialog.show();
-        DatabaseReference mDatabaseReference = FirebaseDatabase.getInstance().getReference().child("schools");
-        final DatabaseReference newPlace = mDatabaseReference.child(model.getId());
+        HashMap<String,Object> map = new HashMap<>();
         if (!TextUtils.isEmpty(name)) {
-            newPlace.child("name").setValue(name);
+            map.put(School.NAME,name);
         }
 
         if (!TextUtils.isEmpty(mail)) {
-            newPlace.child("mail").setValue(mail);
+            map.put(School.MAIL,mail);
         }
 
         if (!TextUtils.isEmpty(website)) {
-            newPlace.child("website").setValue(website);
+            map.put(School.WEBSITE,website);
         }
 
         if (!TextUtils.isEmpty(description)) {
-            newPlace.child("description").setValue(description);
+            map.put(School.DESCRIPTION,description);
         }
 
         if (!TextUtils.isEmpty(mobile)) {
-            newPlace.child("mobileNumber").setValue(mobile);
+            map.put(School.MOBILE_NUMBER,mobile);
         }
 
-        DatabaseReference addressRefrence = newPlace.child("address");
+        HashMap<String,String> address = new HashMap<>();
 
         if (!TextUtils.isEmpty(line1)) {
-            addressRefrence.child("addressLine1").setValue(line1);
+            address.put(School.ADDRESSLINE1,line1);
         }
         if (!TextUtils.isEmpty(line2)) {
-            addressRefrence.child("addressLine2").setValue(line2);
+            address.put(School.ADDRESSLINE2,line2);
         }
         if (!TextUtils.isEmpty(city)) {
-            addressRefrence.child("addressCity").setValue(city);
+            address.put(School.ADDRESS_CITY,city);
         }
         if (!TextUtils.isEmpty(state)) {
-            addressRefrence.child("addressState").setValue(state);
+            address.put(School.ADDRESS_STATE,state);
         }
         if (!TextUtils.isEmpty(country)) {
-            addressRefrence.child("addressCountry").setValue(country);
+            address.put(School.ADDRESS_COUNTRY,country);
         }
         if (str.length == 2) {
             String postalCode = str[1].trim();
-            addressRefrence.child("postalCode").setValue(postalCode);
-
+            address.put(School.POSTAL_CODE,postalCode);
         }
+
+        map.put(School.ADDRESS,address);
 
         if (latLng !=null){
             double latitude = latLng.latitude;
-            newPlace.child("latitude").setValue(latitude);
+            map.put(School.LATITUDE,latitude);
             double longitude = latLng.longitude;
-            newPlace.child("longitude").setValue(longitude);
+            map.put(School.LONGITUDE,longitude);
         }
 
-        mProgressDialog.dismiss();
-        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-        builder.setCancelable(true);
-        builder.setMessage("Congrats, your place updated successfully !");
-        builder.setPositiveButton("Ok",
-                new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog,
-                                        int which) {
-                        fragment_container.setVisibility(View.VISIBLE);
-                        school_edit_home.setVisibility(View.GONE);
-                        dialog.dismiss();
+        FirebaseFirestore.getInstance().collection(School.SCHOOL_DATABASE).document(model.getId()).update(map)
+                .addOnCompleteListener(task -> {
+                    mProgressDialog.dismiss();
+                    if (getActivity()!=null) {
+                        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+                        builder.setCancelable(true);
+                        if (task.isSuccessful()) {
+                            builder.setMessage(R.string.congrats_place_add_msg);
+
+                        } else {
+                            builder.setMessage(R.string.went_wrong_try_later);
+                        }
+                        builder.setPositiveButton(R.string.ok,
+                                (dialog, which) -> {
+                                    fragment_container.setVisibility(View.VISIBLE);
+                                    school_edit_home.setVisibility(View.GONE);
+                                    dialog.dismiss();
+                                });
+                        AlertDialog alert = builder.create();
+                        alert.show();
                     }
                 });
-        AlertDialog alert = builder.create();
-        alert.show();
+
     }
-
-    public void onButtonPressed(Uri uri) {
-        if (mListener != null) {
-            mListener.onFragmentInteraction(uri);
-        }
-    }
-
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        if (context instanceof OnFragmentInteractionListener) {
-            mListener = (OnFragmentInteractionListener) context;
-        } else {
-            throw new RuntimeException(context.toString()
-                    + " must implement OnFragmentInteractionListener");
-        }
-    }
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-        mListener = null;
-    }
-
-
-    public interface OnFragmentInteractionListener {
-        void onFragmentInteraction(Uri uri);
-    }
-
-
 
     @Override
     public void onPause() {
@@ -973,18 +776,17 @@ public class SchoolHomeFragment extends Fragment {
     }
 
     @Override
-    public void onSaveInstanceState(Bundle outState) {
+    public void onSaveInstanceState(@NonNull Bundle outState) {
         super.onSaveInstanceState(outState);
     }
     @Override
-    public void onLowMemory()
-    {
+    public void onLowMemory() {
         super.onLowMemory();
     }
     @Override
     public void onResume() {
         super.onResume();
-        ConnectionUtil.checkConnection(getActivity().findViewById(R.id.placeSnackBar));
+        if (getActivity()!=null)ConnectionUtil.checkConnection(getActivity().findViewById(R.id.placeSnackBar));
     }
     @Override
     public void onDestroyView() {
