@@ -1,63 +1,31 @@
 package com.credolabs.justcredo.school;
 
-import android.app.Activity;
-import android.app.ActivityOptions;
-import android.content.Intent;
-import android.graphics.Color;
-import android.graphics.PorterDuff;
 import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
-import android.os.Handler;
-import android.support.annotation.RequiresApi;
-import android.support.design.widget.AppBarLayout;
-import android.support.design.widget.CollapsingToolbarLayout;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.design.widget.TabLayout;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.AppCompatImageView;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.WindowManager;
 import android.widget.LinearLayout;
 import android.widget.ProgressBar;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import com.credolabs.justcredo.FullZoomImageViewActivity;
 import com.credolabs.justcredo.R;
 import com.credolabs.justcredo.adapters.ViewPagerAdapter;
 import com.credolabs.justcredo.internet.ConnectionUtil;
 import com.credolabs.justcredo.model.School;
 import com.credolabs.justcredo.model.ZoomObject;
-import com.credolabs.justcredo.newplace.BoardsFragment;
-import com.credolabs.justcredo.newplace.PlaceExtraFragment;
-import com.credolabs.justcredo.newplace.PlaceFacilitiesFragment;
 import com.credolabs.justcredo.newplace.PlaceTypes;
-import com.credolabs.justcredo.newplace.TypePlaceFragment;
-import com.credolabs.justcredo.sliderlayout.CirclePageIndicator;
-import com.credolabs.justcredo.sliderlayout.ImageSlideAdapter;
-import com.credolabs.justcredo.utility.CustomeToastFragment;
 import com.credolabs.justcredo.utility.Util;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
+import com.google.firebase.firestore.FirebaseFirestore;
 
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Timer;
-import java.util.TimerTask;
 
-public class SchoolDetailActivity extends AppCompatActivity implements
-                SchoolHomeFragment.OnFragmentInteractionListener,SchoolFacilitiesFragment.OnFragmentInteractionListener,
+public class SchoolDetailActivity extends AppCompatActivity implements SchoolFacilitiesFragment.OnFragmentInteractionListener,
                 SchoolAddressFragment.OnFragmentInteractionListener, SchoolGalleryFragment.OnFragmentInteractionListener,
                 SchoolBlogsFragment.OnFragmentInteractionListener,SchoolReviewFragment.OnFragmentInteractionListener{
 
@@ -73,28 +41,27 @@ public class SchoolDetailActivity extends AppCompatActivity implements
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_school_detail);
-        final Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        progress = (ProgressBar)findViewById(R.id.progress);
-        not_found = (LinearLayout) findViewById(R.id.not_found);
-        final TextView not_found_text1 = (TextView) findViewById(R.id.not_found_text1);
-        not_found_text1.setText("Something went wrong.");
+        final Toolbar toolbar = findViewById(R.id.toolbar);
+        progress =findViewById(R.id.progress);
+        not_found = findViewById(R.id.not_found);
+        final TextView not_found_text1 = findViewById(R.id.not_found_text1);
+        not_found_text1.setText(R.string.something_went_wrong);
         final TextView not_found_text2 = (TextView) findViewById(R.id.not_found_text2);
-        not_found_text2.setText("Sorry ! Please try again later.");
+        not_found_text2.setText(R.string.try_again_later);
         not_found.setVisibility(View.GONE);
         setSupportActionBar(toolbar);
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        if (getSupportActionBar()!=null) getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         getSupportActionBar().setDisplayShowHomeEnabled(true);
         ConnectionUtil.checkConnection(findViewById(R.id.placeSnackBar));
-        String id = getIntent().getStringExtra("SchoolDetail");
-        DatabaseReference schoolReference = FirebaseDatabase.getInstance().getReference().child("schools");
-        schoolReference.child(id).addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                model = dataSnapshot.getValue(School.class);
+        String id = getIntent().getStringExtra(School.ID);
+        FirebaseFirestore.getInstance().collection(School.SCHOOL_DATABASE).document(id).get().addOnCompleteListener(task -> {
+            progress.setVisibility(View.GONE);
+            if (task.isSuccessful() && task.getResult() !=null){
+                model = task.getResult().toObject(School.class);
                 if (model != null) {
                     toolbar.setTitle(model.getName());
                     if (model.getImages()!=null) {
-                        imagesList = new ArrayList<String>(model.getImages().values());
+                        imagesList = new ArrayList<>(model.getImages().values());
                         String address = Util.getAddress(model.getAddress());
                         zoomObject.setImages(imagesList);
                         zoomObject.setName(model.getName());
@@ -104,37 +71,30 @@ public class SchoolDetailActivity extends AppCompatActivity implements
                     }else {
                         imagesList = new ArrayList<>();
                     }
-                    tabLayout = (TabLayout) findViewById(R.id.profile_tablayout);
-                    viewPager = (ViewPager) findViewById(R.id.profile_viewPager);
+                    tabLayout = findViewById(R.id.profile_tablayout);
+                    viewPager = findViewById(R.id.profile_viewPager);
                     ViewPagerAdapter adapter = new ViewPagerAdapter(getSupportFragmentManager());
-                    adapter.addFragment(SchoolHomeFragment.newInstance(model,model.getName()), "Details"); //Main
+                    adapter.addFragment(SchoolHomeFragment.newInstance(model), getString(R.string.details));                           //Main
                     if (model.getType().equalsIgnoreCase(PlaceTypes.MUSIC.getValue())){
-                        adapter.addFragment(SchoolGenreFragment.newInstance(model,model.getName()),"Genres");
+                        adapter.addFragment(SchoolGenreFragment.newInstance(model,model.getName()),getString(R.string.genres));
                     }
                     if (model.getType()!=null &&(model.getType().equalsIgnoreCase(PlaceTypes.ART.getValue())||
                             model.getType().equalsIgnoreCase(PlaceTypes.SPORTS.getValue())||
                             model.getType().equalsIgnoreCase(PlaceTypes.COACHING.getValue()))){
-                        adapter.addFragment(SchoolClassesTypeFragment.newInstance(model,model.getName()),"Classes");
+                        adapter.addFragment(SchoolClassesTypeFragment.newInstance(model,model.getName()),getString(R.string.classes));
                     }
-                    adapter.addFragment(SchoolFacilitiesFragment.newInstance(model,model.getName()), "Facilities"); //Post
-                    adapter.addFragment(SchoolGalleryFragment.newInstance(imagesList,model.getUserID(),model.getId()), "Gallery");//ProfileBookmarksFragment
-                    adapter.addFragment(SchoolBlogsFragment.newInstance(model,model.getName()), "Blogs");//ProfileBookmarksFragment
-                    adapter.addFragment(SchoolReviewFragment.newInstance(model.getId(),"schools"), "Reviews");//ProfileBookmarksFragment
-                    adapter.addFragment(SchoolAddressFragment.newInstance(model,model.getName()), "Contact");//ProfileBookmarksFragment
-
-
+                    adapter.addFragment(SchoolFacilitiesFragment.newInstance(model,model.getName()), getString(R.string.facilities_heading));           //Post
+                    adapter.addFragment(SchoolGalleryFragment.newInstance(imagesList,model.getUserID(),model.getId()), getString(R.string.gallery));    //ProfileBookmarksFragment
+                    adapter.addFragment(SchoolBlogsFragment.newInstance(model,model.getName()), getString(R.string.blogs));                             //ProfileBookmarksFragment
+                    adapter.addFragment(SchoolReviewFragment.newInstance(model.getId(),School.SCHOOL_DATABASE), getString(R.string.reviews));           //ProfileBookmarksFragment
+                    adapter.addFragment(SchoolAddressFragment.newInstance(model,model.getName()), getString(R.string.contact));                         //ProfileBookmarksFragment
                     viewPager.setAdapter(adapter);
                     tabLayout.setupWithViewPager(viewPager);
-                    //viewPager.setOffscreenPageLimit(6);
-                    progress.setVisibility(View.GONE);
+                }else {
+                    not_found.setVisibility(View.VISIBLE);
                 }
-            }
-
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-                progress.setVisibility(View.GONE);
+            }else {
                 not_found.setVisibility(View.VISIBLE);
-
             }
         });
     }
